requirements/base.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
lxml==4.9.2
pyTelegramBotAPI==4.10.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.14
